Add column controller
